requirements.txt
Flask==2.3.3
PyGitHub==2.3.0
requests==2.31.0
google-generativeai==0.3.0  # Adjust version based on Gemini API availability